Add files via upload
Create an application that runs in any modern web browser in which you can simulate controlling a robot. The robot will be placed on a grid of 5x5. Supply controls that let you move the robot forwards in the direction it is facing. Through the same controls you can rotate the robot in place to face any cardinal direction. Use any Javascript framework you’re comfortable with.